breastcancer-main (1)/breastcancer-main/requirements.txt
fastapi
fastapi-sqlalchemy
pydantic
alembic
psycopg2
uvicorn
python-dotenv
pandas
Jinja2==3.0.3
python-multipart
numpy
matplotlib
tensorflow
torchvision
h5py
google-cloud-bigquery
db-dtypes
google-cloud-storage
